+++ src/main/java/duke/Launcher.java
@@ -3,6 +3,7 @@ package duke;
 import javafx.application.Application;
 
 public class Launcher {
+
     public static void main(String[] args) {
         Application.launch(Main.class, args);
     }
