--- src/main/java/duke/Ui.java
@@ -1,19 +1,0 @@
-package duke;
-
-/**
- * Ui is a class that encapsulates the attributes and behaviours of a User Interface.
- *
- * @author leezhixuan
- */
-public class Ui {
-    private String name;
-
-    /**
-     * Creates an instance of Ui.
-     *
-     * @param name Name of chat bot.
-     */
-    public Ui(String name) {
-        this.name = name;
-    }
-}
