+++ src/main/java/duke/command/ByeCommand.java
@@ -1,9 +1,8 @@
 package duke.command;
 
 import duke.Duke;
-import duke.Storage;
-import duke.ToDoList;
-import duke.Ui;
+import duke.util.Storage;
+import duke.util.ToDoList;
 
 /**
  * ByeCommand is a Command that encapsulates the attributes and behaviour of a command to stop.
@@ -11,7 +10,6 @@ import duke.Ui;
 public class ByeCommand extends Command {
 
     private Duke chatBot;
-    private Ui ui;
     private ToDoList tdl;
     private Storage storage;
 
@@ -19,13 +17,11 @@ public class ByeCommand extends Command {
      * Creates an instance of ByeCommand.
      *
      * @param chatBot The instance of chat bot itself.
-     * @param ui Instance of User Interface currently in use.
      * @param tdl Instance of ToDoList currently in use.
      * @param storage Instance of Storage currently in use.
      */
-    public ByeCommand(Duke chatBot, Ui ui, ToDoList tdl, Storage storage) {
+    public ByeCommand(Duke chatBot, ToDoList tdl, Storage storage) {
         this.chatBot = chatBot;
-        this.ui = ui;
         this.tdl = tdl;
         this.storage = storage;
     }
