+++ src/main/java/duke/command/DeleteCommand.java
@@ -1,7 +1,6 @@
 package duke.command;
 
-import duke.ToDoList;
-import duke.Ui;
+import duke.util.ToDoList;
 
 /**
  * DeleteCommand is a Command that encapsulates the attributes and behavior of deleting a Task from the
@@ -11,19 +10,16 @@ import duke.Ui;
  */
 public class DeleteCommand extends Command {
     private ToDoList tdl;
-    private Ui ui;
     private int index;
 
     /**
      * Creates an instance of DeleteCommand.
      *
      * @param tdl Instance of ToDoList in use.
-     * @param ui Instance of User Interfaces in use.
      * @param index Index of the Task on the ToDoList that is to be deleted.
      */
-    public DeleteCommand(ToDoList tdl, Ui ui, int index) {
+    public DeleteCommand(ToDoList tdl, int index) {
         this.tdl = tdl;
-        this.ui = ui;
         this.index = index;
     }
