+++ src/main/java/duke/command/ToDoCommand.java
@@ -1,6 +1,6 @@
 package duke.command;
 
-import duke.ToDoList;
+import duke.util.ToDoList;
 
 /**
  * ToDoCommand is a Command that encapsulates the attributes and behaviours of adding a ToDo to ToDoList.
