+++ src/main/java/duke/util/ToDoList.java
@@ -1,4 +1,4 @@
-package duke;
+package duke.util;
 
 import java.time.LocalDateTime;
 import java.util.ArrayList;
@@ -31,6 +31,8 @@ public class ToDoList {
 
     /**
      * Displays the content in ToDoList as a numbered list.
+     *
+     * @return A String representation of the TodoList.
      */
     public String displayList() {
         Integer number = 1;
@@ -50,11 +52,12 @@ public class ToDoList {
      * Marks the item at index num on the ToDoList as done.
      *
      * @param num The index of the item that is to be marked as done.
+     * @return A success message to indicate the Task has been marked as done.
      */
     public String markAsDone(int num) {
         this.record.get(num - 1).setCompleted();
         String response = "Good! Good! Took you long enough to complete this: \n";
-        response += "  " + "[X] " + record.get(num - 1).toString() +"\n";
+        response += "  " + "[X] " + record.get(num - 1).toString() + "\n";
         return response;
     }
 
@@ -62,6 +65,7 @@ public class ToDoList {
      * Adds a ToDo Task to the ToDoList.
      *
      * @param item Name of ToDo.
+     * @return A success message indicating that a ToDo has been added to the ToDoList.
      */
     public String addToDo(String item) {
         ToDo todo = new ToDo(item);
@@ -78,6 +82,7 @@ public class ToDoList {
      *
      * @param item Name of Event.
      * @param duration Time between start and end of Event.
+     * @return A success message indicating that an Event has been added to the ToDoList.
      */
     public String addEvent(String item, String duration) {
         Event event = new Event(item, duration);
@@ -94,13 +99,14 @@ public class ToDoList {
      *
      * @param item Name of Task.
      * @param deadline Date and Time at which the Task is due.
+     * @return A success message indicating that a Deadline has been added to the ToDoList.
      */
     public String addDeadline(String item, LocalDateTime deadline) {
         Deadline dl = new Deadline(item, deadline);
         this.record.add(dl);
         this.totalNumber++;
         String response = "Aye Aye Capt'. I've added this Deadline:\n";
-        response += "  [D][ ] " + dl.toString() +"\n";
+        response += "  [D][ ] " + dl.toString() + "\n";
         response += "Now you have " + this.totalNumber.toString() + " in the list";
         return response;
     }
@@ -109,6 +115,7 @@ public class ToDoList {
      * Deletes the Task at index on the ToDoList.
      *
      * @param index Index of the Task to be deleted from ToDoList.
+     * @return A success message indicating that the Task-to-be-deleted has been deleted from ToDoList.
      */
     public String delete(int index) {
         Task removed = this.record.remove(index - 1);
@@ -144,6 +151,7 @@ public class ToDoList {
      * Looks for the Task which has a name that contains the specified string
      *
      * @param target String to look out for in the Task's name
+     * @return A String representation of the List of items that matched.
      */
     public String find(String target) {
         ArrayList<Task> temp = new ArrayList<>();
@@ -158,9 +166,9 @@ public class ToDoList {
             response += "Here are the matching tasks in your list: \n";
             for (Task t : temp) {
                 if (t.isCompleted()) {
-                    response += number.toString() + "." + t.logo() + "[X] " + t.toString() +"\n";
+                    response += number.toString() + "." + t.logo() + "[X] " + t.toString() + "\n";
                 } else {
-                    response += number.toString() + "." + t.logo() + "[ ] " + t.toString() +"\n";
+                    response += number.toString() + "." + t.logo() + "[ ] " + t.toString() + "\n";
                 }
                 number++;
             }
@@ -170,4 +178,3 @@ public class ToDoList {
         return response;
     }
 }
-
