+++ src/test/java/duke/command/DoneCommandTest.java
@@ -4,8 +4,7 @@ import static org.junit.jupiter.api.Assertions.fail;
 
 import org.junit.jupiter.api.Test;
 
-import duke.ToDoList;
-import duke.Ui;
+import duke.util.ToDoList;
 
 
 public class DoneCommandTest {
@@ -13,9 +12,8 @@ public class DoneCommandTest {
     @Test
     public void execute_indexOutOfBounds_exceptionThrown() {
         String name = "Testing Bot";
-        Ui ui = new Ui(name);
         ToDoList tdl = new ToDoList(name);
-        DoneCommand dc = new DoneCommand(tdl, ui, 100);
+        DoneCommand dc = new DoneCommand(tdl, 100);
         try {
             dc.execute();
         } catch (IndexOutOfBoundsException e) {
